+++ applications/fishing-map/src/features/debug/debug.slice.ts
@@ -3,6 +3,7 @@ import { RootState } from 'store'
 
 export enum DebugOption {
   Blob = 'blob',
+  PresenceTrackPOC = 'presenceTrackPOC',
   Extruded = 'extruded',
   Debug = 'debug',
   Thinning = 'thinning',
@@ -19,9 +20,10 @@ const initialState: DebugState = {
   active: false,
   options: {
     blob: false,
-    extruded: false,
     debug: false,
     thinning: true,
+    extruded: false,
+    presenceTrackPOC: false,
   },
 }