+++ applications/fishing-map/src/features/resources/resources.selectors.ts
@@ -1,7 +1,8 @@
 import { createSelector } from '@reduxjs/toolkit'
-import { resolveDataviewsResourceQueries } from '@globalfishingwatch/dataviews-client'
-import { selectDataviewInstancesResolved } from 'features/dataviews/dataviews.selectors'
 import { selectVisibleEvents } from 'features/app/app.selectors'
+import { ThinningLevels, THINNING_LEVELS } from 'data/config'
+import { selectDebugOptions } from 'features/debug/debug.slice'
+import { isGuestUser } from 'features/user/user.selectors'
 import { selectResources } from './resources.slice'
 
 export const selectVisibleResources = createSelector(
@@ -20,7 +21,13 @@ export const selectVisibleResources = createSelector(
   }
 )
 
-export const selectDataviewsResourceQueries = createSelector(
-  [selectDataviewInstancesResolved],
-  (dataviewInstances) => resolveDataviewsResourceQueries(dataviewInstances ?? [])
+export const selectThinningConfig = createSelector(
+  [(state) => isGuestUser(state), selectDebugOptions],
+  (guestUser, { thinning }) => {
+    if (!thinning) return null
+    const thinningConfig = guestUser
+      ? THINNING_LEVELS[ThinningLevels.Aggressive]
+      : THINNING_LEVELS[ThinningLevels.Default]
+    return thinningConfig
+  }
 )