+++ applications/fishing-map/src/features/workspace/vessels/VesselLayerPanel.tsx
@@ -2,7 +2,12 @@ import React, { Fragment, useState } from 'react'
 import cx from 'classnames'
 import { useSelector } from 'react-redux'
 import { Trans, useTranslation } from 'react-i18next'
-import { DatasetTypes, ResourceStatus, Vessel } from '@globalfishingwatch/api-types'
+import {
+  Vessel,
+  DatasetTypes,
+  ResourceStatus,
+  DataviewDatasetConfigParam,
+} from '@globalfishingwatch/api-types'
 import { IconButton, Tooltip } from '@globalfishingwatch/ui-components'
 import { ColorBarOption } from '@globalfishingwatch/ui-components/dist/color-bar'
 import { Segment } from '@globalfishingwatch/data-transforms'
@@ -67,7 +72,9 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
 
   const vesselLabel = infoResource?.data ? getVesselLabel(infoResource.data) : ''
   const vesselId =
-    (infoResource?.datasetConfig?.params?.find((p) => p.id === 'vesselId')?.value as string) ||
+    (infoResource?.datasetConfig?.params?.find(
+      (p: DataviewDatasetConfigParam) => p.id === 'vesselId'
+    )?.value as string) ||
     dataview.id.replace(VESSEL_DATAVIEW_INSTANCE_PREFIX, '') ||
     ''
   const vesselTitle = vesselLabel || vesselId