+++ applications/vessel-history/src/App.tsx
@@ -1,23 +1,18 @@
 import React, { Fragment, useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
-import GFWAPI from '@globalfishingwatch/api-client'
-import useGFWLogin from '@globalfishingwatch/react-hooks/dist/use-login'
 import Home from 'features/home/Home'
 import { getLocationType } from 'routes/routes.selectors'
 import { PROFILE, SETTINGS } from 'routes/routes'
-import { BASE_URL, SPLASH_TIMEOUT } from 'data/constants'
+import { SPLASH_TIMEOUT } from 'data/constants'
 import Profile from 'features/profile/Profile'
 import Splash from 'features/splash/Splash'
 import Settings from 'features/settings/Settings'
+import { useUser } from 'features/user/user.hooks'
 import './App.css'
 
 function App() {
-  const { loading, logged } = useGFWLogin(GFWAPI)
+  const { loading, logged, authorized } = useUser()
   const [minLoading, setMinLoading] = useState(true)
-  if (!loading && !logged) {
-    const location = window.location.origin + BASE_URL
-    window.location.assign(GFWAPI.getLoginUrl(location))
-  }
 
   // Splash screen is shown at least one second
   useEffect(() => {
@@ -27,8 +22,8 @@ function App() {
 
   const locationType = useSelector(getLocationType)
 
-  if (loading || minLoading) {
-    return <Splash />
+  if (loading || minLoading || !logged || !authorized) {
+    return <Splash intro={minLoading} />
   }
   if (locationType === SETTINGS) {
     return <Settings />