+++ applications/vessel-history/src/features/profile/components/activity/Activity.tsx
@@ -3,6 +3,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import AutoSizer from 'react-virtualized-auto-sizer'
 import { VariableSizeList as List } from 'react-window'
 import { Modal, Spinner } from '@globalfishingwatch/ui-components'
+import { ApiEvent } from '@globalfishingwatch/api-types/dist'
 import { selectResourcesLoading } from 'features/resources/resources.slice'
 import { VesselWithHistory } from 'types'
 import {
@@ -12,6 +13,8 @@ import {
 import { fetchRegionsThunk } from 'features/regions/regions.slice'
 import ActivityFilters from 'features/profile/filters/ActivityFilters'
 import { fetchPsmaThunk } from 'features/psma/psma.slice'
+import { setHighlightedEvent } from 'features/map/map.slice'
+import { useLocationConnect } from 'routes/routes.hook'
 import ActivityItem from './ActivityItem'
 import ActivityModalContent from './ActivityModalContent'
 import styles from './Activity.module.css'
@@ -19,6 +22,7 @@ interface ActivityProps {
   vessel: VesselWithHistory | null
   lastPosition: any
   lastPortVisit: any
+  onMoveToMap: () => void
 }
 
 const Activity: React.FC<ActivityProps> = (props): React.ReactElement => {
@@ -35,6 +39,13 @@ const Activity: React.FC<ActivityProps> = (props): React.ReactElement => {
   }, [])
   const closeModal = useCallback(() => setIsOpen(false), [])
 
+  const { dispatchQueryParams } = useLocationConnect()
+  const selectEventOnMap = useCallback((event: RenderedEvent) => {
+    dispatch(setHighlightedEvent({id: event.id} as ApiEvent))
+    dispatchQueryParams({ latitude: event.position.lat, longitude: event.position.lon })
+    props.onMoveToMap()
+  }, [dispatch, dispatchQueryParams, props])
+
   useEffect(() => {
     dispatch(fetchRegionsThunk())
     dispatch(fetchPsmaThunk())
@@ -50,7 +61,7 @@ const Activity: React.FC<ActivityProps> = (props): React.ReactElement => {
             {selectedEvent && <ActivityModalContent event={selectedEvent}></ActivityModalContent>}
           </Modal>
           <div className={styles.activityContainer}>
-            {events && events.length > 0 && (
+            {events && events.length > 0 ? (
               <AutoSizer disableWidth={true}>
                 {({ width, height }) => (
                   <List
@@ -67,6 +78,7 @@ const Activity: React.FC<ActivityProps> = (props): React.ReactElement => {
                           <ActivityItem
                             key={index}
                             event={event}
+                            onMapClick={(event) => selectEventOnMap(event)}
                             onInfoClick={(event) => openModal(event)}
                           />
                         </div>
@@ -75,6 +87,8 @@ const Activity: React.FC<ActivityProps> = (props): React.ReactElement => {
                   </List>
                 )}
               </AutoSizer>
+            ) : (
+              <p>No events results, try change the filters</p>
             )}
           </div>
         </Fragment>