+++ applications/vessel-history/src/features/search/search.thunk.ts
@@ -1,19 +1,71 @@
 import { createAsyncThunk } from '@reduxjs/toolkit'
-import GFWAPI from '@globalfishingwatch/api-client'
+import { stringify } from 'qs'
+import GFWAPI, {
+  AdvancedSearchQueryField,
+  getAdvancedSearchQuery,
+} from '@globalfishingwatch/api-client'
 import { VesselSearch } from '@globalfishingwatch/api-types'
 import { BASE_DATASET, RESULTS_PER_PAGE, SEARCH_MIN_CHARACTERS } from 'data/constants'
 import { RootState } from 'store'
 import { CachedVesselSearch } from './search.slice'
 
-const fetchData = async (query: string, offset: number, signal?: AbortSignal | null) => {
-  return await GFWAPI.fetch<any>(
-    `/v1/vessels/search?datasets=${encodeURIComponent(
-      BASE_DATASET
-    )}&limit=${RESULTS_PER_PAGE}&offset=${offset}&query=${encodeURIComponent(query)}&useTMT=true`,
+export const getSerializedQuery = (query: string, advancedSearch?: Record<string, any>) => {
+  if (!advancedSearch) return query
+  const fields: AdvancedSearchQueryField[] = [
     {
-      signal,
-    }
-  )
+      key: 'shipname',
+      value: query,
+    },
+    {
+      key: 'mmsi',
+      value: advancedSearch.mmsi,
+    },
+    {
+      key: 'imo',
+      value: advancedSearch.imo,
+    },
+    {
+      key: 'callsign',
+      value: advancedSearch.callsign,
+    },
+    {
+      key: 'flag',
+      value: advancedSearch.flags.map((f: string) => ({ id: f })),
+    },
+    {
+      key: 'lastTransmissionDate',
+      value: advancedSearch.lastTransmissionDate,
+    },
+    {
+      key: 'firstTransmissionDate',
+      value: advancedSearch.firstTransmissionDate,
+    },
+  ]
+  return getAdvancedSearchQuery(fields)
+}
+
+const fetchData = async (
+  query: string,
+  offset: number,
+  signal?: AbortSignal | null,
+  advancedSearch?: Record<string, any>
+) => {
+  const serializedQuery = getSerializedQuery(query, advancedSearch)
+  const endpoint = advancedSearch ? 'advanced-search' : 'search'
+
+  const urlQuery = stringify({
+    datasets: BASE_DATASET,
+    limit: RESULTS_PER_PAGE,
+    offset,
+    query: serializedQuery,
+    useTMT: true,
+  })
+
+  const url = `/v1/vessels/${endpoint}?${urlQuery}`
+
+  return await GFWAPI.fetch<any>(url, {
+    signal,
+  })
     .then((json: any) => {
       const resultVessels: Array<VesselSearch> = json.entries
 
@@ -29,12 +81,12 @@ const fetchData = async (query: string, offset: number, signal?: AbortSignal | n
       return null
     })
 }
-const searchNeedsFetch = (
+const getSearchNeedsFetch = (
   query: string,
   offset: number,
   metadata: CachedVesselSearch | null
 ): boolean => {
-  if (query.length <= SEARCH_MIN_CHARACTERS) {
+  if (query.length < SEARCH_MIN_CHARACTERS) {
     return false
   }
   if (!metadata) {
@@ -56,20 +108,25 @@ const searchNeedsFetch = (
 export type VesselSearchThunk = {
   query: string
   offset: number
+  advancedSearch?: Record<string, any>
 }
 
 export const fetchVesselSearchThunk = createAsyncThunk(
   'search/vessels',
-  async ({ query, offset }: VesselSearchThunk, { rejectWithValue, getState, signal }) => {
-    const searchData = await fetchData(query, offset, signal)
+  async (
+    { query, offset, advancedSearch }: VesselSearchThunk,
+    { rejectWithValue, getState, signal }
+  ) => {
+    const searchData = await fetchData(query, offset, signal, advancedSearch)
     return searchData
   },
   {
-    condition: ({ query, offset }, { getState, extra }) => {
+    condition: ({ query, offset, advancedSearch }, { getState, extra }) => {
+      const serializedQuery = getSerializedQuery(query, advancedSearch)
       const { search } = getState() as RootState
-      const metadata: CachedVesselSearch = search.queries[query]
-
-      return searchNeedsFetch(query, offset, metadata)
+      const metadata: CachedVesselSearch = search.queries[serializedQuery]
+      const searchNeedsFetch = getSearchNeedsFetch(serializedQuery, offset, metadata)
+      return searchNeedsFetch
     },
   }
 )