+++ applications/vessel-history/src/features/splash/Splash.test.tsx
@@ -1,10 +1,61 @@
-import React from 'react'
-import renderer from 'react-test-renderer'
+import { render, waitFor } from '@testing-library/react'
+import { useTranslation } from 'react-i18next'
+import { useUser } from 'features/user/user.hooks'
 import Splash from './Splash'
 
+jest.mock('features/user/user.hooks')
+jest.mock('react-i18next')
+
 describe('Splash', () => {
-  it('renders correctly', () => {
-    const tree = renderer.create(<Splash />).toJSON()
-    expect(tree).toMatchSnapshot()
+  const mockUseTranslation: jest.Mock = useTranslation as jest.Mock
+  const mockUseUser: jest.Mock = useUser as jest.Mock
+  const userDefault = {
+    loading: true,
+    logged: false,
+    user: null,
+    error: null,
+    authorized: null,
+    login: () => jest.fn(),
+    logout: () => jest.fn(),
+  }
+
+  beforeEach(() => {
+    mockUseTranslation.mockReturnValue({
+      t: (_ns: any, fallback: any) => fallback,
+    })
+  })
+
+  afterEach(() => {
+    jest.clearAllMocks()
+  })
+
+  it('renders correctly for auth user', async () => {
+    const user = { ...userDefault, logged: true, authorized: true, user: { id: '1', name: 'foo' } }
+    mockUseUser.mockReturnValue(user)
+
+    const component = render(<Splash />)
+    const splashElement = await waitFor(() => component.getByTestId('splash'))
+    expect(splashElement).toMatchSnapshot()
+    expect(splashElement).toBeInTheDocument()
+  })
+
+  it('renders not logged in message', async () => {
+    const user = { ...userDefault }
+    mockUseUser.mockReturnValue(user)
+
+    const component = render(<Splash />)
+    const splashElement = await waitFor(() => component.getByTestId('splash'))
+    expect(splashElement).toMatchSnapshot()
+    expect(splashElement).toBeInTheDocument()
+  })
+
+  it('renders not authorized in message', async () => {
+    const user = { ...userDefault, logged: true, authorized: false, user: { id: '1', name: 'foo' } }
+    mockUseUser.mockReturnValue(user)
+
+    const component = render(<Splash />)
+    const splashElement = await waitFor(() => component.getByTestId('splash'))
+    expect(splashElement).toMatchSnapshot()
+    expect(splashElement).toBeInTheDocument()
   })
 })