+++ packages/layer-composer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/layer-composer
 
+## 9.2.2
+
+### Patch Changes
+
+- fix typo
+
+## 9.2.1
+
+### Patch Changes
+
+- 28d7d17c: fix missing encounter vessel name
+
 ## 9.2.0
 
 ### Minor Changes