+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @globalfishingwatch/ui-components
 
+## 11.0.1
+
+### Patch Changes
+
+- 74932493: allow switch id to use with htmlFor
+
+## 11.0.0
+
+### Patch Changes
+
+- 25440a9f: fix disabled select option
+  - @globalfishingwatch/react-hooks@14.0.0
+
 ## 10.2.0
 
 ### Minor Changes